Made Some Changes to the service

+++ doStuff/Services/Service.cs
@@ -146,86 +146,98 @@ namespace doStuff.Services
         #region FriendRelations
         public bool AnswerFriendRequest(int userId, int senderId, bool answer)
         {
-
-            if (db.ExistsUserToUserRelation(senderId, userId) || (db.ExistsUserToUserRelation(userId, senderId)))
+            UserToUserRelation relation = db.GetUserToUserRelation(senderId, userId);
+            if(relation != null && relation.Active)
             {
-                UserToUserRelation relation = db.GetUserToUserRelation(senderId, userId);
                 relation.Answer = answer;
-                relation.Active = true;
                 return db.SetUserToUserRelation(relation);
             }
-            else
-            {
-                return false;
-            }
-
+            return false;
         }
         public bool RemoveFriend(int userId, int friendId)
         {
-
-            if (!db.ExistsUserToUserRelation(userId, friendId) && (!db.ExistsUserToUserRelation(friendId, userId)))
-            {
-                throw new UserNotFoundException();
-            }
-            else
+            if (IsFriendsWith(userId, friendId))
             {
                 UserToUserRelation relation = db.GetUserToUserRelation(userId, friendId);
+                if (relation == null || relation.Active == false)
+                {
+                    relation = db.GetUserToUserRelation(friendId, userId);
+                }
+                relation.Answer = false;
+                return db.SetUserToUserRelation(relation);
             }
-
-            return false;
+            //TODO REMOVE IF STATEMENT
+            throw Exception("You Tried To Remove A Friend Without Checking IsFriendsWith(userId, friendId) In The Controller First!!!");
         }
         public bool SendFriendRequest(int userId, int friendId)
         {
-
-            //TODO Check if user has already sent a request before.
             if (!db.ExistsUserToUserRelation(userId, friendId))
             {
                 UserToUserRelation relation = new UserToUserRelation();
+                relation.Active = true;
                 relation.SenderId = userId;
                 relation.ReceiverId = friendId;
+                relation.Answer = null;
                 return db.CreateUserToUserRelation(relation);
             }
             else
             {
                 return false;
             }
-
         }
         #endregion
         #region GroupRelations
         public bool AddMember(int userId, int groupId)
         {
-            return false;// db.CreateGroupToUserRelation(groupId, userId);
+            if (IsOwnerOfGroup(userId, groupId))
+            {
+                GroupToUserRelation relation = db.GetGroupToUserRelation(groupId, userId);
+                if (relation == null)
+                {
+                    relation = new GroupToUserRelation();
+                    relation.Active = true;
+                    relation.GroupId = groupId;
+                    relation.MemberId = userId;
+                    return db.CreateGroupToUserRelation(relation);
+                }
+                relation.Active = true;
+                return db.SetGroupToUserRelation(relation);
+            }
+            throw Exception("You Tried To Add A Member Without Checking IsOwnerOfGroup(userId, groupId) In The Controller First!!!");
         }
         public bool RemoveMember(int userId, int groupId)
         {
-            //TODO: Throw User Exception.
-            GroupToUserRelation relation = db.GetGroupToUserRelation(groupId, userId);
-
-            if (relation == null)
+            if (IsOwnerOfGroup(userId, groupId))
             {
-                throw new UserNotFoundException();
+                GroupToUserRelation relation = db.GetGroupToUserRelation(groupId, userId);
+
+                if (relation == null)
+                {
+                    return false;
+                }
+                return db.RemoveGroupToUserRelation(relation.GroupToUserRelationID);
             }
-            return db.RemoveGroupToUserRelation(relation.GroupToUserRelationID);
+            throw Exception("You Tried To Remove A Member Without Checking IsOwnerOfGroup(userId, groupId) In The Controller First!!!");
         }
         #endregion
         #region EventRelation
         public bool AnswerEvent(int userId, int eventId, bool answer)
         {
-            //TODO: Throw Event Exception.
-            if (db.ExistsEventToUserRelation(eventId, userId))
+            if (IsInvitedToEvent(userId, eventId))
             {
                 EventToUserRelation relation = db.GetEventToUserRelation(eventId, userId);
                 if (relation == null)
                 {
-                    throw new EventNotFoundException();
+                    relation = new EventToUserRelation();
+                    relation.EventId = eventId;
+                    relation.AttendeeId = userId;
+                    relation.Answer = answer;
+                    return db.CreateEventToUserRelation(relation);
                 }
-
                 relation.Answer = answer;
                 return db.SetEventToUserRelation(relation);
             }
-
-            return false;
+            throw Exception("You Tried To Answer An Event Without Checking IsInvitedToEvent(userId, eventId) In The Controller First!!!");
         }
         #endregion
         #region Create
@@ -235,12 +247,8 @@ namespace doStuff.Services
         }
         public bool CreateGroup(Group group)
         {
-            //TODO make user join group automatically
-            bool created = false;
-
-            created = db.CreateGroup(group);
-
-            if (created)
+            //TODO
+            if (db.CreateGroup(group))
             {
                 GroupToUserRelation relation = new GroupToUserRelation();
                 relation.GroupId = group.GroupID;
@@ -330,12 +338,12 @@ namespace doStuff.Services
         }
         #endregion
         #region GetViewModel
-        public EventFeedViewModel GetGroupFeed(int groupId, int userId)
+        public GroupFeedViewModel GetGroupFeed(int groupId, int userId)
         {
             //TODO
             // Show something if user has no friends or events?
 
-            EventFeedViewModel feed = new EventFeedViewModel();
+            GroupFeedViewModel feed = new GroupFeedViewModel();
             List<EventViewModel> eventViews = new List<EventViewModel>();
             List<Event> events = db.GetEvents(groupId);
 
@@ -358,6 +366,7 @@ namespace doStuff.Services
             sidebar.UserList = db.GetMembers(groupId);
             feed.Events = eventViews;
             feed.SideBar = sidebar;
+            feed.groupId = groupId;
 
 
             List<Group> groups = db.GetGroups(userId);
